+++ store_spanner.go
